make hybrid fleet x area
Having an issue when a fleet fishes > 1 area. RN I'm getting F_i by weighting the tuned F from the N_weights, but when this feeds back into Baranov the values are deflated because the resultant Z_is, which figure into the denominator, are distinct. Need to discuss with AEP; I'm not even positive the N weights was a legit thing to do.