prelim figure for Xijas
basically, the values in this matrix don't change; what changes is the AGE at which the matrix values become non-zero depending on the sex and location of the individual. So, X-ijas will look like this for all fish over age 8 (when everyone is guaranteed to be >400mm) but vary beneath that.