re-run output on files saved in 10-10 folder
I thought to do this because there was conflict/redundancies between what is in the output folder for 10/09 and 10/10. The obj/dat/opt in 10/09 say niter44 but are otherwise empty; the fits in that same folder look great as do the ones inside 10/10. However when I run the plots on the opt/dat in 10/10 the fits are junk. So I am not positive which run/config was used to make those nice 40y fits in the first place, only that it involved a cpp ca. 10/09., which notably had no omega_0ij influencing recruitment and an incorrect summation on N_weight.