introduce Neqn and mult by R0 vector
based on mtg with AEP, it was wrong to only use the first column of the solve(I-Mat2) object, because that implies all recruitment happens in area 1 (which explains why C1 was like e+108). Instead we do the full solve(I-Mat2)*R0, which needs to be setup to have a # first, followed by a bunch of zeroes, then a # again for the next area.