age typo in mortality; revert to correct half-death sequence
the correct order of events is that they are born; the midyear NAA are those that move and experience half of natM; then this midyear NAA is fished (which is why there's no need for a mortality term in the catch prediction calc); then the end biomass is the midyear biomass minus catches and after a second death round. The NAA at the beginning of next year is the same as the end of this year, plus rec.