save fixed partial mortality - abandon v3L
post meeting w AEP i realized the negative NAA are not because of h/M weirdness; it's because the continuous F as written in the hybrid method can be >1, whereas the way instF is implemented inthe NAA equation is actually setup to expect less than 1. hence negs. Will change how F is calculated, then build in bi-annual catches.