remove last rogue mla_yais multiplier
I am pretty sure this was supposed to be a placeholder for the LAA matrix, which would make it so that we were getting expected numbers-at-length given the growth in the year at hand. In turn the second mla_yais (used in the power function to get weight) should be the length_yais_mid or whatever; the mla_yais is really a placeholder for all of the last few months' of runs with length stuff turned off.