add 0.2 to surv_yf_error to match document
This is the francis 2011 approach. We have small sds from vast which are year and fleet specific. We want to give the model a little more flexibility. For example, the same obs + pred with a lower SD will return a higher likelihood, and vice versa. So hopefully by increasing the sigma here we will be able to reduce runtime.