input correct survey error values into fit plot
these will change OTF with whatever df$survey_yf_err is, but need to pay attention to how they are treated. Per Jim's email we are convinced the SD_log values from VAST, which I'm passing to the NLL, are real SDs (in log space). Thus for plotting we should bring them in to normal space. I think this just means we exponentiate them.