revert growth to three stages
this is represented on p32,36 of notebook.
basically, we need to retain knowledge of what's happening to the mean LAA year to year. This means that we indeed need to calculate LAA_beg based on LAA_end for years 1 onwards (year zero gets straight vonB treatment). Thus the 0.5 growth stages make sense; one half between BEG and MID (concurrent with movement reweight) and once more between MID and END. The plus group reweighting is only applicable to the END to BEG(y+1) transition since it deals with fish of age PLUS and PLUS-1 aging up one year.