turn on ans_sdr likelihood
right now b-y is a vector of length yRun.
AEP and I discussed how since WC is likely not at equilibrium we'd probably want a stock specific b penalty for that region, which in this case would be very low (since exp(small number) = small number, thus penalizing the expected recruitment in that area at the beginning).

That said, almost all areas are way too high from the outset, so my first tests will look at whether the SDR likelihood with b_y as-is is enough to get in the ballpark.