quick run with slx fixed and bc vast truncated to 2003
recall that the index working with here is still "damaged" in the respect that oddly no data before 2003 was included in the stdization for BC, when there is definitely BC_std values before then. I'm also thinking that using vast & bc_early is double dipping coverage because bc_early is in fact bc_std. so i think we either need to run with bc_vast, having coverage from 1980-2019 (real data 1991-2019) OR use bc_early (bc_std) and bc strs as two separate fleets, with no vast in question. this run is a compromise between teh two where in bc-early is in there and bc vast is truncated to start in 2003, but note that 2003-2010 is made up of BOTH std and strs data (thus double dipping).