us surv_yf_obs as mean for simulate
these were returning zero cause I was using the predicted values for future years. Now it will give a randomized value for every single year of forecast based on the 5-year mean provided. this might change a little with seed but otherwise not much by estimation since the input survey error and means are fixed.