turned on full array for future catch, surv
might not be necessary - was getting eigen error bc the loop thru tEnd fails when looking for obs(yr_future). one way around this is to fill the observed with zeros, so it can proceed to simulation. easier way is to leave them as-is and have the flag proceed if the obs is either -1 (aka NA) or if we are in a forecast year. then be sure to return a value for forecast years.