bundle upload - slow and NaN errors
somehow between the successful test run on 2/15 and the changes made here, i started getting nan evals when calling nlminb() and obj() takes forever (20mins) to construct. In lieu of waiting 20 mins x N permutations to diagnose if this is a problem in df, mappy, or bounds, I'm pushing these and then going to revert to what the repo looked like when I pushed the successful run with acomps input. the good thing is the cpp hasn't changed; it's clearly a data formatting issue.