catchcomp like on, survcomp like off
I noticed that GOA SURV has NSAMP == NA. Tracked down to a0 slx for that fleet is NaN but not super clear why that would happen (for just one age). the blks are correct for that fleet.

Confirmed that the cpp will throw Nas if either comp fleet is on, because the NaN in nsamp cascades into the total likelihood.